requirements.txt
Flask
Flask-Login
Werkzeug
Pillow
Flask-Mail
Flask-SQLAlchemy
Flask-Bcrypt
gunicorn